BioForm/requirements.txt
asgiref==3.5.2
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
cffi==1.15.1
cryptography==38.0.1
distlib==0.3.5
Django==4.1
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
filelock==3.8.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==1.7.1
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
virtualenv==20.16.3